requirements.txt
streamlit
joblib
numpy
Pillow
requests
scikit-learn
